+++ src/main/java/com/basf/challenge/http/ResponseHandler.java
@@ -1,4 +1,4 @@
-package com.basf.challenge.service;
+package com.basf.challenge.http;
 
 import com.basf.challenge.dto.ResponseDataDTO;
 import com.basf.challenge.dto.ResponseMessageDTO;
