+++ src/test/java/com/basf/challenge/repository/PatentRepositoryTests.java
@@ -1,7 +1,6 @@
-package com.basf.challenge;
+package com.basf.challenge.repository;
 
 import com.basf.challenge.entity.Patent;
-import com.basf.challenge.repository.PatentRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
