+++ src/test/java/com/basf/challenge/services/ChallengeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.basf.challenge;
+package com.basf.challenge.services;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
