+++ src/test/java/com/basf/challenge/services/ReadXMLPropertiesTest.java
@@ -1,4 +1,4 @@
-package com.basf.challenge;
+package com.basf.challenge.services;
 
 import com.basf.challenge.dto.Book;
 import com.basf.challenge.dto.xmlProperties.QuestelPatentDocument;
